Move 'externalPush' to jdk8
